app/requirements.txt
Flask==0.12.0
Flask-Babel
markdown
PyICU
